+++ MaxMind.MinFraud/Response/Address.cs
@@ -20,8 +20,7 @@ namespace MaxMind.MinFraud.Response
         /// This property is <c>true</c> if the postal code
         /// provided with the address is in the city for the address.The property is
         /// <c>false</c> when the postal code is not in the city. If the address was
-        /// not provided, could not be parsed, or was not in USA, the property will
-	/// be <c>null</c>.
+        /// not provided or could not be parsed, the property will be <c>null</c>.
         /// </summary>
         [JsonProperty("is_postal_in_city")]
         public bool? IsPostalInCity { get; internal set; }